requirements.txt
django==1.10
celery
djangorestframework==3.4.0  
psycopg2==2.5.3  
redis==2.10.5
movie_info_downloader